use int instead of long for sublime text 3 compatibility

+++ goto_last_edit.py
@@ -6,13 +6,13 @@ LAST_EDITS_SETTING = 'last_edits'
 
 class RecordLastEdit(sublime_plugin.EventListener):
     def on_modified(self, view):
-        last_edits = view.settings().get(LAST_EDITS_SETTING, {}) 
-        edit_position = view.sel()[0] 
+        last_edits = view.settings().get(LAST_EDITS_SETTING, {})
+        edit_position = view.sel()[0]
         last_edits[str(view.id())] = {'a': edit_position.a, 'b': edit_position.b}
-        view.settings().set(LAST_EDITS_SETTING, last_edits) 
+        view.settings().set(LAST_EDITS_SETTING, last_edits)
 
- 
-class GotoLastEdit(sublime_plugin.TextCommand): 
+
+class GotoLastEdit(sublime_plugin.TextCommand):
     # The position the cursor was at before the command fired. Saved when the
     # command is run, so that if the user runs the command again before making
     # another edit in the file, the cursor returns to its original position.
@@ -21,12 +21,12 @@ class GotoLastEdit(sublime_plugin.TextCommand):
     def move_cursor_to_region(self, region):
         """ Clear the cursor's position and move it to `region`. """
         cursor = self.view.sel()
-        self.original_position = cursor[0] 
+        self.original_position = cursor[0]
         cursor.clear()
         cursor.add(region)
-        self.view.show(region) 
+        self.view.show(region)
 
-    def run(self, edit): 
+    def run(self, edit):
         """
         If there was a last edit recorded for the view, store the current
         position as self.original_position and move the cursor to the position
@@ -35,7 +35,7 @@ class GotoLastEdit(sublime_plugin.TextCommand):
         If the cursor is currently at the same position as the last edit, and
         there `self.original_position` is available, then return the cursor to
         its original position.
-        """ 
+        """
         last_edits = self.view.settings().get(LAST_EDITS_SETTING, {})
         last_edit = last_edits.get(str(self.view.id()), None)
         current_position = self.view.sel()[0]
@@ -44,7 +44,7 @@ class GotoLastEdit(sublime_plugin.TextCommand):
             return
 
         last_edit_region = sublime.Region(
-            long(last_edit['a']), long(last_edit['b']))
+            int(last_edit['a']), int(last_edit['b']))
 
         if self.original_position is not None \
             and current_position == last_edit_region:
